Move clap_cache_dir() to maple_core
