+++ crates/cli/src/command/cache.rs
@@ -1,10 +1,11 @@
 use anyhow::Result;
 use clap::{Parser, Subcommand};
 use maple_core::datastore::CACHE_INFO_IN_MEMORY;
+use maple_core::dirs::clap_cache_dir;
 use std::fs::read_dir;
 use std::io::Write;
 use std::path::{PathBuf, MAIN_SEPARATOR};
-use utils::{clap_cache_dir, remove_dir_contents};
+use utils::remove_dir_contents;
 
 /// List and remove all the cached contents.
 #[derive(Subcommand, Debug, Clone)]
