+++ crates/maple_core/src/cache/mod.rs
@@ -96,7 +96,7 @@ impl CacheInfo {
                 && now.signed_duration_since(digest.last_visit).num_days() < MAX_DAYS
                 // In case the cache was not created completely.
                 && std::fs::File::open(&digest.cached_path)
-                    .and_then(crate::utils::count_lines)
+                    .and_then(utils::count_lines)
                     .map(|total| total == digest.total)
                     .unwrap_or(false)
             {
@@ -191,7 +191,7 @@ pub fn store_cache_digest(
     shell_cmd: ShellCommand,
     new_created_cache: PathBuf,
 ) -> std::io::Result<Digest> {
-    let total = crate::utils::count_lines(std::fs::File::open(&new_created_cache)?)?;
+    let total = utils::count_lines(std::fs::File::open(&new_created_cache)?)?;
 
     let digest = Digest::new(shell_cmd, total, new_created_cache);
